Event System implemented again
Instead of just checking for no enemies on map when we DecrementActiveLandCount(), I have moved the changeGamestate(BIOME_SELECTION) to EventManager, basically EventManager now handles when to go to BiomeSelection (Event is cleared). Modded WorldManager: to get a get function for ActiveLandCount and remove the previous wave clear system as mentioned. Modded EnemySpawner.cs: Made the L key on press set EventClearStatus to true (Debugging tool). Modded BiomeSelectUI: setsEventClearStatus to false on card pop up (to not be stuck in a forever card popup loop)